# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAPhoto API
  description: |-
    Each user will be presented with a stream of photos in reverse chronological order, with information
    about when it was uploaded (date and time) and how many likes and comments it has. Users can place
    (and later remove) a “like” to photos from other users. Also, users can add comments to any photo
    (even those uploaded by themself). Comments can be removed by the author only.

    Users can ban other users. If Alice (user) bans Eve (user), Eve won’t be able to see any information
    about Alice. Alice can decide to remove the ban at any moment.

    Users will have their own personal profiles. The personal profile page for user detail: user’s photos (in
    reverse chronological order), their followers and following, and how many photos have been uploaded.
    Each user can change his/her own username, upload photos, remove photos, and follow/unfollow
    other users. Removal of a photo will also remove likes and comments.

    Users profiles can be searched via username. The login is perfomed by specifying the username.
  contact: 
    name: Simone La Bella 
    email: labella.1995847@studenti.uniroma1.it
  version: 1.0.0
...
components:
  schemas:
...
    commentRequest:
      title: comment
      description: "Comment schema"
      type: object
      properties: 
        photoOwner:
          description:  Owner of the photo.
          type: string
          pattern: '^.*?$'
          example: 56A
          minLength: 3
          maxLength: 16
        photoID:
          description: Unique photo identifier.
          type: string
          pattern: '^.*?$'
          example: 472S
          minLength: 3
          maxLength: 16
        userID:
          description: Unique user identifier.
          type: string
          pattern: '^.*?$'
          example: 56A
          minLength: 3
          maxLength: 16
        content:
          description: Content of the comment
          type: string
          example: "Hi to everyone"
          pattern: '^.*?$'
          minLength: 3
          maxLength: 16
    commentResponse:
      title: comment
      description: "Comment schema"
      type: object
      properties: 
        photoOwner:
          description:  Owner of the photo.
          type: string
          pattern: '^.*?$'
          example: 56A
          readOnly: true  
          minLength: 3
          maxLength: 16
        photoID:
          description: Unique photo identifier.
          type: string
          pattern: '^.*?$'
          example: 472S
          readOnly: true
          minLength: 3
          maxLength: 16
        commentID:
          description: "Comment ID"
          type: string
          pattern: '^.*?$'
          example: JKW
          readOnly: true
          minLength: 3
          maxLength: 16
        userID:
          description: Unique user identifier.
          type: string
          pattern: '^.*?$'
          example: 56A
          minLength: 3
          maxLength: 16
          readOnly: true
        content:
          description: Content of the comment
          type: string
          example: "Hi to everyone"
          minLength: 3
          maxLength: 16
    follow:
      title: follow
      description: "Follow schema"
      type: object
      properties: 
        followID:
          description: "Follow ID"
          type: string
          pattern: '^.*?$'
          example: JJKS
          readOnly: true
          minLength: 3
          maxLength: 16
        userID:
          description: Unique user identifier.
          type: string
          pattern: '^.*?$'
          example: 56A
          minLength: 3
          maxLength: 16
          readOnly: true
        userFollowedID:
          description: Unique user identifier.
          type: string
          pattern: '^.*?$'
          example: 47A
          minLength: 3
          maxLength: 16
          readOnly: true
    ban:
      title: ban
      description: "ban schema"
      type: object
      properties: 
        banID:
          description: "Ban ID"
          type: string
          pattern: '^.*?$'
          example: BB7
          readOnly: true
          minLength: 3
          maxLength: 16
        userID:
          description: Unique user identifier.
          type: string
          pattern: '^.*?$'
          example: 56A
          minLength: 3
          maxLength: 16
          readOnly: true
        userBannedID:
          description: Unique user identifier.
          type: string
          pattern: '^.*?$'
          example: 47A
          minLength: 3
          maxLength: 16
          readOnly: true    
...
  parameters:
    username:
      schema:
          description: username schema.
          type: string
          pattern: '^.*?$'
          example: MarioRossi
          minLength: 5
          maxLength: 30
      name: username
      in: path
      required: true
      description: "Username's user."
    photoIDPathParam:
      schema:
        description: photo schema.
        type: string
        pattern: '^.*?$'
        example: JK2
        readOnly: true
        minLength: 3
        maxLength: 15
      name: photoID
      in: path
      required: true
      description: "Unique ID of a specific photo"
...
    commentIDPathParam:
      schema:
        description: "Comment ID"
        type: string
        pattern: '^.*?$'
        example: KIS8
        readOnly: true
        minLength: 3
        maxLength: 16
      name: commentID
      in: path
      required: true
      description: "Unique ID of a specific comment"
...
    followID:
      schema:
        description: "Follow ID"
        type: string
        pattern: '^.*?$'
        example: JSL3
        readOnly: true
        minLength: 3
        maxLength: 16
      name: followID
      in: path
      required: true
      description: "Unique ID of a specific follow"
    banID:
      schema:
        description: "Ban ID"
        type: string
        pattern: '^.*?$'
        example: JD33
        readOnly: true
        minLength: 3
        maxLength: 16
      name: banID
      in: path
      required: true
      description: "Unique ID of a specific ban"
  responses:
    BadRequest:
      description: The server cannot or will not process the request due to something that is perceived to be a client error.
    InternalServerError:
      description:  the server encountered an unexpected condition that prevented it from fulfilling the request.
    UnauthorizedError:
      description: The client request has not been completed because it lacks valid authentication credentials for the requested resource
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
paths:
...
  /user/{username}/photo/{photoID}/comment/{commentID}:
    parameters:
      - $ref : "#/components/parameters/username"
      - $ref : "#/components/parameters/photoIDPathParam"
      - $ref : "#/components/parameters/commentIDPathParam"

    delete: 
      operationId: uncommentPhoto
      summary: Uncomment a photo
      description: Remove comment from the system
      security:
          - bearerAuth : []
      tags: ["Comment"]
      responses:
        "204":
          description: Comment removed correctly
        "402":
          description: Comment not found
          
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "500": { $ref: "#/components/responses/InternalServerError" } 
  /user/{username}/photo/{photoID}/comment:
    parameters: 
      - $ref: "#/components/parameters/username"
      - $ref : "#/components/parameters/photoIDPathParam"

    post: 
      operationId: commentPhoto
      summary: Comment a photo 
      tags: ["Comment"]
      security:
          - bearerAuth : []
      description: Add comment in the system
      requestBody:
        content: 
          application/json:
            schema: { $ref: '#/components/schemas/commentRequest'}
      responses:
        "201":
          description: Comment created succesfully
          content:
            application/json:
              schema: { $ref: '#/components/schemas/commentResponse'}
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
          

    get:
      operationId: getComment
      tags: ["Comment"]
      security:
          - bearerAuth : []
      summary: Return Comment
      description: |
        Return comment of a specific photo.
      responses:
        "200":
          description: Comment found correctly.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/commentResponse'}
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "500": { $ref: "#/components/responses/InternalServerError" } 
  /user/{username}/follow/{followID}:
    parameters:
      - $ref : "#/components/parameters/username"
      - $ref : "#/components/parameters/followID"

    delete:
      operationId: unfollowUser
      summary: Unollow User 
      tags: ["Follow"]
      security:
          - bearerAuth : []
      description: Remove Follow from the system
      responses:
        "204":
          description: Follow removed correctly
        "402":
          description: Follow not found
          
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "500": { $ref: "#/components/responses/InternalServerError" }   
  /user/{username}/follow:

    parameters: 
      - $ref : "#/components/parameters/username"
        
    post: 
      operationId: followUser
      summary: Follow User 
      tags: ["Follow"]
      security:
          - bearerAuth : []
      description: Add Follow in the system
      responses:
        "201":
          description: Follow added correctly
          content:
            application/json:
              schema: { $ref: "#/components/schemas/follow"}
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "500": { $ref: "#/components/responses/InternalServerError" }

    get:
      operationId: getFollow
      tags: ["Follow"]
      security:
        - bearerAuth : []
      summary: Return follow
      description: |
        Return the follow.
      responses:
        "200":
          description: Follow founded correctly.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/follow'}
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "500": { $ref: "#/components/responses/InternalServerError" }  
  /user/{username}/ban/{banID}:
    parameters:
      - $ref : "#/components/parameters/username"
      - $ref : '#/components/parameters/banID'
    delete:
      operationId: unbanUser
      summary: Unban User 
      tags: ["Ban"]
      security:
          - bearerAuth : []
      description: Remove Ban from the system
      responses:
        "204":
          description: Ban removed correctly
        "402":
          description: Ban not found
          
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "500": { $ref: "#/components/responses/InternalServerError" }  
  /user/{username}/ban:
    parameters: 
      - $ref : "#/components/parameters/username"
    post: 
      operationId: banUser
      summary: Ban User 
      tags: ["Ban"]
      security:
          - bearerAuth : []
      description: Add Ban in the system
      responses:
        "201":
          description: Ban executed correctly
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ban"}
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "500": { $ref: "#/components/responses/InternalServerError" }

    get:
      operationId: getBan
      tags: ["Ban"]
      security:
        - bearerAuth : []
      summary: Return Ban
      description: |
        Return the ban founded.
      responses:
        "200":
          description: Ban founded correctly.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ban'}
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "500": { $ref: "#/components/responses/InternalServerError" }   
...